pyproject.toml
[project]
name = "abe-froman"
version = "0.1.0"
description = "Workflow orchestrator using LangGraph and Claude Agent SDK"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "langgraph>=1.0",
]

[project.scripts]
abe-froman = "abe_froman.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/abe_froman"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
